Got denoising to work kinda (read notes)
I got the wave file to come out and play, however rather than filtering
only it is also speeding up the sound and making you sound like a
chipmunk.

i'm assuming this is due to the fact that the filter function is
truncating the array, however my attempt at fixing this with the
'return 0' in the filter function did not alleviate this.